+++ README.md
@@ -15,6 +15,13 @@
 
 Excelize is a library written in pure Go and providing a set of functions that allow you to write to and read from XLSX files. Support reads and writes XLSX file generated by Microsoft Excel&trade; 2007 and later. Support save file without losing original charts of XLSX. This library needs Go version 1.8 or later. The full API docs can be seen using go's built-in documentation tool, or online at [godoc.org](https://godoc.org/github.com/360EntSecGroup-Skylar/excelize) and [docs reference](https://xuri.me/excelize/).
 
+**WARNING!**
+
+Please note that row manipulation methods using different row numbering.
+Some part of methods using zero-based row indexses while another using Excel row numbering starting with 1.
+
+Please use Row methods with suffix `2` (`RemoveRow2()` for example) if present to operate Excel row numberinng.   
+
 ## Basic Usage
 
 ### Installation
